Bugfix: Change float to long. Result was negative.
git-svn-id: https://vcs.informatik.uni-siegen.de/svn/tpml/trunk@861 69bfd797-8304-0410-886d-9c04947c740f

+++ pgtest3/source/de/unisiegen/tpml/Optimizer.java
@@ -138,7 +138,7 @@ public final class Optimizer
       {
         end = 1 ;
       }
-      double last = 0 ;
+      long last = 0 ;
       final DecimalFormat df1 = new DecimalFormat ( "00" ) ;
       final DecimalFormat df2 = new DecimalFormat ( "00,000 ms" ) ;
       final DecimalFormat df3 = new DecimalFormat ( "00.0" ) ;
